check: extract validation running to another function
This is so that we can reuse this logic later on
