+++ check/check.js
@@ -1,5 +1,7 @@
-const get = require('lodash.get');
 const validator = require('validator');
+
+const runner = require('./runner');
+
 const extraValidators = ['contains', 'equals', 'matches'];
 
 module.exports = (fields, locations) => {
@@ -7,42 +9,8 @@ module.exports = (fields, locations) => {
   fields = Array.isArray(fields) ? fields : [fields];
 
   const middleware = (req, res, next) => {
-    const validationErrors = [];
-    const allFields = [];
-
-    fields.forEach(field => locations.forEach(location => {
-      allFields.push({
-        location,
-        path: field,
-        value: location === 'headers' ? req.get(field) : get(req[location], field)
-      });
-    }));
-
-    const promises = allFields.filter(field => {
-      return locations.length > 1 ? field.value !== undefined : true;
-    }).map(field => {
-      return validators.reduce((promise, validatorCfg) => promise.then(() => {
-        const result = validatorCfg.custom ?
-          validatorCfg.validator(field.value, req) :
-          validatorCfg.validator(String(field.value), ...validatorCfg.options);
-
-        return Promise.resolve(result).then(result => {
-          if (!result) {
-            throw new Error('Invalid value');
-          }
-        });
-      }).catch(err => {
-        validationErrors.push({
-          location: field.location,
-          path: field.path,
-          value: field.value,
-          message: validatorCfg.message || err.message
-        });
-      }), Promise.resolve());
-    });
-
-    return Promise.all(promises).then(() => {
-      req._validationErrors = (req._validationErrors || []).concat(validationErrors);
+    return runner(req, middleware._context).then(errors => {
+      req._validationErrors = (req._validationErrors || []).concat(errors);
       next();
     }, next);
   };
@@ -72,6 +40,12 @@ module.exports = (fields, locations) => {
   middleware.withMessage = message => {
     validators[validators.length - 1].message = message;
     return middleware;
+  };
+
+  middleware._context = {
+    fields,
+    locations,
+    validators
   };
 
   return middleware;
